services/api/requirements.txt
alembic==1.14.1
fastapi==0.115.8
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
sqlalchemy==2.0.38
uvicorn[standard]==0.34.0

